Ensure NCCL_BLOCKING_WAIT=1 works for dist.barrier()
Blocking wait didn't work for dist.barrier() since we performed a
cudaDeviceSynchronize() before we performed any of the timeout checks. As a
result, in case of failures/desync the barrier() call would get stuck on
cudaDeviceSynchrnonize() and would never return a timeout error to the user.

To fix this, I've moved the device synchronization after the timeout checks.

Differential Revision: [D22108899](https://our.internmc.facebook.com/intern/diff/D22108899/)

ghstack-source-id: 106123004
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40207